Refactor: 내가 개설한 CoffeeChat 목록 조회시 CustomPageResponse를 반환하도록 수정 - CustomPageInfo의 이름을 CustomPageResponse로 수정하고 매개변수를 단순화 했습니다. - 서비스와 컨트롤러에서 별도의 Dto로 또 감싸지 않고 CustomPageResponse를 반환하도록 했습니다. 그에 따라 불필요한 dto를 삭제했습니다.
