+++ module-api/src/main/java/kernel/jdon/coffeechat/service/CoffeeChatService.java
@@ -12,12 +12,11 @@ import kernel.jdon.coffeechat.dto.response.CreateCoffeeChatResponse;
 import kernel.jdon.coffeechat.dto.response.DeleteCoffeeChatResponse;
 import kernel.jdon.coffeechat.dto.response.FindCoffeeChatListResponse;
 import kernel.jdon.coffeechat.dto.response.FindCoffeeChatResponse;
-import kernel.jdon.coffeechat.dto.response.HostCoffeeChatListResponse;
 import kernel.jdon.coffeechat.dto.response.UpdateCoffeeChatResponse;
 import kernel.jdon.coffeechat.error.CoffeeChatErrorCode;
 import kernel.jdon.coffeechat.repository.CoffeeChatRepository;
 import kernel.jdon.global.exception.ApiException;
-import kernel.jdon.global.page.CustomPageInfo;
+import kernel.jdon.global.page.CustomPageResponse;
 import kernel.jdon.member.domain.Member;
 import kernel.jdon.member.error.MemberErrorCode;
 import kernel.jdon.member.repository.MemberRepository;
@@ -51,17 +50,12 @@ public class CoffeeChatService {
 		coffeeChat.increaseViewCount();
 	}
 
-	public HostCoffeeChatListResponse findHostCoffeeChatList(Long memberId, Pageable pageable) {
-		Page<FindCoffeeChatListResponse> findCoffeeChatListResponse = coffeeChatRepository.findAllByMemberIdAndIsDeletedFalse(
-				memberId,
-				pageable)
+	public CustomPageResponse<FindCoffeeChatListResponse> findHostCoffeeChatList(Long memberId, Pageable pageable) {
+		Page<FindCoffeeChatListResponse> findCoffeeChatPage = coffeeChatRepository.findAllByMemberIdAndIsDeletedFalse(
+				memberId, pageable)
 			.map(FindCoffeeChatListResponse::of);
 
-		CustomPageInfo<FindCoffeeChatListResponse> customPageInfo = new CustomPageInfo<>(
-			findCoffeeChatListResponse.getContent(), pageable,
-			findCoffeeChatListResponse.getTotalElements());
-
-		return HostCoffeeChatListResponse.of(customPageInfo);
+		return new CustomPageResponse<>(findCoffeeChatPage);
 	}
 
 	@Transactional
